feat(frontend): update the app to use newer versions and fix imports
