+++ app/src/components/resetGridButton.tsx
@@ -1,6 +1,6 @@
-import React, { useContext } from "react";
-import { Button } from "@material-ui/core";
-import { GridContext } from "../providers/GridContext";
+import React, { useContext } from 'react';
+import { Button } from '@mui/material';
+import { GridContext } from '../providers/GridContext';
 
 export const ResetGridButton: React.FC = () => {
   const { resetGrid } = useContext(GridContext);
@@ -9,7 +9,7 @@ export const ResetGridButton: React.FC = () => {
     resetGrid();
   };
   return (
-    <div style={{ padding: "5px" }}>
+    <div style={{ padding: '5px' }}>
       <Button onClick={handleClick}>Reset Grid</Button>
     </div>
   );
