--- frontend/src/constants.tsx
@@ -1,32 +1,0 @@
-export const DEFAULT_GRID_SIZE = 20;
-export const MIN_GRID_SIZE = 5;
-export const MAX_GRID_SIZE = 30;
-
-export const GAME_OF_LIFE_URI = "http://localhost:8080/game-of-life";
-export const GAME_OF_LIFE_MULTITHREAD_URI =
-  "http://localhost:8080/game-of-life-multithreaded";
-export const STRUCTURE_URI = "http://localhost:8080/structures";
-
-export const DEAD = 0;
-export const ALIVE = 1;
-export const MUTANT = 2;
-
-export const NUM_TYPES = 3;
-
-export const TYPES = [
-  "none",
-  "block",
-  "beehive",
-  "loaf",
-  "bread",
-  "tub",
-  "blinker",
-  "toad",
-  "beacon",
-  "pulsar",
-  "pentadecathlon",
-  "glider",
-  "light-weight-spaceship",
-  "middle-weight-spaceship",
-  "large-weight-spaceship",
-];
